+++ packages/controls/js/libs/box-spacing-control/components/margin/margin-bottom.js
@@ -108,6 +108,7 @@ export function MarginBottom({
 			if (!openPopover) setFocusSide('');
 			setLabelClassName('');
 		},
+		threshold: 0,
 	});
 
 	if (marginDisable === 'all' || marginDisable === 'vertical') {