+++ packages/controls/js/libs/box-spacing-control/components/margin/margin-left.js
@@ -109,6 +109,7 @@ export function MarginLeft({
 			if (!openPopover) setFocusSide('');
 			setLabelClassName('');
 		},
+		threshold: 0,
 	});
 
 	if (marginDisable === 'all' || marginDisable === 'horizontal') {