+++ packages/controls/js/libs/box-spacing-control/components/padding/padding-bottom.js
@@ -110,6 +110,7 @@ export function PaddingBottom({
 			if (!openPopover) setFocusSide('');
 			setLabelClassName('');
 		},
+		threshold: 0,
 	});
 
 	if (paddingDisable === 'all' || paddingDisable === 'vertical') {